+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v3.4.3 - February 6th, 2026
 ### Fixed
 - Gift Button removing InvisibleTypingButton and Send Button
+- Add Reaction Button not removing properly
 
 ## v3.4.2 - February 5th, 2026
 ### Changed